DT-500 Fix printing in gnome-text-editor
Gnome-text-editor crashed when trying to print. This is because
Gtk can't find the modules. The first patch binded the modules
folder into the root, but James commented that it could be
fixed just with an environment variable.

This patch sets that variable for both GTK4 and GTK2, which
were both missing (it was set only for GTK3).